Updated pallet constraint calculation

+++ primitives/statement-store/src/runtime_api.rs
@@ -31,9 +31,9 @@ use sp_externalities::ExternalitiesExt;
 #[derive(Clone, PartialEq, Eq, Encode, Decode, RuntimeDebug, TypeInfo)]
 pub struct ValidStatement {
 	/// Max statement count for this account, as calculated by the runtime.
-	pub max_count: u64,
+	pub max_count: u32,
 	/// Max total data size for this account, as calculated by the runtime.
-	pub max_size: u64,
+	pub max_size: u32,
 	/// Global priority value. This is used to prioritize statements on the global scale.
 	/// If the global loimit of messages is reached, the statement with the lowest priority will be
 	/// removed first.
